Rearrange Files and Widgets
Put the form back together, attempt to carry alertTime through code to database. Delete unused calendar functions.
