+++ lib/page/todo.dart
@@ -1,15 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:circular_check_box/circular_check_box.dart';
-import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:provider/provider.dart';
 import 'package:todo_dot/model/task.dart';
-import 'package:todo_dot/page/datepicker.dart';
 import 'package:todo_dot/widget/todo_list.dart';
 import 'package:todo_dot/widget/taskform.dart';
 import 'package:todo_dot/database/database.dart';
-import 'package:todo_dot/database/utils.dart';
 import 'sidebar.dart';
-import 'editTask.dart';
 import 'package:todo_dot/style.dart';
 
 //https://pub.dev/packages/circular_check_box
@@ -103,6 +98,7 @@ class _AddTaskState extends State<AddTask> {
   final _formKey = GlobalKey<FormState>();
   String title = '';
   String description = '';
+  DateTime alertTime;
 
   @override
   Widget build(BuildContext context) => AlertDialog(
@@ -121,15 +117,11 @@ class _AddTaskState extends State<AddTask> {
               onChangedTitle: (title) => setState(() => this.title = title),
               onChangedDescription: (description) =>
                   setState(() => this.description = description),
-            ),
-            SizedBox(height: 10),
-            DatetimePickerWidget(),
-            SizedBox(height: 30),
-            TaskFormButtons(
+              onChangedAlert: (alertTime) =>
+                  setState(() => this.alertTime = alertTime),
               onSavedTask: addTask,
-              onDeleteTask: cancelTask,
+              onCancel: cancelTask,
             ),
-
           ],
         ),
       ));
@@ -143,6 +135,7 @@ class _AddTaskState extends State<AddTask> {
         id: DateTime.now().toString(),
         title: title,
         description: description,
+        alertTime: alertTime,
         createdTime: DateTime.now(),
       );
 
@@ -152,120 +145,12 @@ class _AddTaskState extends State<AddTask> {
       Navigator.of(context).pop();
     }
   }
-}
 
-void cancelTask() {
-  // task never created, return to list page
-  return;
-}
-
-class TaskWidget extends StatelessWidget {
-  final Task task;
-
-  const TaskWidget({
-    @required this.task,
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => ClipRRect(
-        borderRadius: BorderRadius.circular(25),
-        child: Slidable(
-          actionPane: SlidableDrawerActionPane(),
-          key: Key(task.id),
-          actions: [
-            IconSlideAction(
-              color: Colors.green,
-              onTap: () => editTask(context, task),
-              caption: 'Edit',
-              icon: Icons.edit,
-            )
-          ],
-          secondaryActions: [
-            IconSlideAction(
-              color: Colors.red,
-              caption: 'Delete',
-              onTap: () => deleteTask(context, task),
-              icon: Icons.delete,
-            ),
-          ],
-          child: buildTask(context),
-        ),
-      );
-
-  Widget buildTask(BuildContext context) => Container(
-        color: Colors.white,
-        padding: EdgeInsets.all(20),
-        child: Row(
-          children: [
-            CircularCheckBox(
-              value: task.isComplete,
-              checkColor: Colors.white,
-              activeColor: todoMediumGreen,
-              inactiveColor: Colors.redAccent,
-              disabledColor: Colors.grey,
-              onChanged: (_) {
-                final provider =
-                    Provider.of<TaskProvider>(context, listen: false);
-                final isDone = provider.toggleCompleteStatus(task);
-
-                Utils.showSnackBar(
-                  context,
-                  isDone ? 'Task completed' : 'Task marked incomplete',
-                );
-              },
-            ),
-            SizedBox(
-              width: 20,
-            ),
-            Expanded(
-                child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(
-                  task.title,
-                  style: TextStyle(
-                    fontSize: 28,
-                  ),
-                ),
-                if (task.description.isNotEmpty)
-                  Container(
-                    margin: EdgeInsets.only(top: 4),
-                    child: Text(
-                      task.description,
-                      style: TextStyle(
-                        fontSize: 20,
-                      ),
-                    ),
-                  ),
-                if (task.alertTime != null)
-                  Container(
-                    margin: EdgeInsets.only(top: 4),
-                    child: Text(
-                      task.alertTime.toLocal().toString(),
-                      style: TextStyle(
-                        fontSize: 20,
-                      ),
-                    ),
-                  ),
-              ],
-            ))
-          ],
-        ),
-      );
-
-  void deleteTask(BuildContext context, Task task) {
-    final provider = Provider.of<TaskProvider>(context, listen: false);
-    provider.removeTask(task);
-
-    Utils.showSnackBar(context, 'Deleted the task');
+  void cancelTask() {
+    // task never created, return to list page
+    Navigator.of(context).pop();
+    return;
   }
-
-  void editTask(BuildContext context, Task task) => Navigator.of(context).push(
-        MaterialPageRoute(
-          builder: (context) => EditTaskPage(task: task),
-        ),
-      );
 }
 
 Widget buildText(String text) => Center(
