+++ lib/widget/taskform.dart
@@ -1,20 +1,27 @@
 import 'package:flutter/material.dart';
+import 'package:todo_dot/page/datepicker.dart';
 import 'package:todo_dot/style.dart';
 
 class TaskForm extends StatelessWidget {
   final String title;
   final String description;
-  final DateTime alarmTime;
+  final DateTime alertTime;
   final ValueChanged<String> onChangedTitle;
   final ValueChanged<String> onChangedDescription;
+  final ValueChanged<DateTime> onChangedAlert;
+  final VoidCallback onSavedTask;
+  final VoidCallback onCancel;
 
   const TaskForm({
     Key key,
     this.title = '',
     this.description = '',
-    this.alarmTime,
+    this.alertTime,
     @required this.onChangedTitle,
     @required this.onChangedDescription,
+    @required this.onChangedAlert,
+    @required this.onSavedTask,
+    @required this.onCancel,
   }) : super(key: key);
 
   @override
@@ -25,7 +32,16 @@ class TaskForm extends StatelessWidget {
             buildTitle(),
             SizedBox(height: 8),
             buildDescription(),
+            SizedBox(height: 10),
+            DatetimePickerWidget(),
             SizedBox(height: 25),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: [
+                buildSaveButton(),
+                buildCancelButton(),
+              ],
+            )
           ],
         ),
       );
@@ -56,71 +72,45 @@ class TaskForm extends StatelessWidget {
         ),
       );
 
-}
-
-
-class TaskFormButtons extends StatelessWidget {
-  final VoidCallback onSavedTask;
-  final VoidCallback onDeleteTask;
-
-  const TaskFormButtons({
-    Key key,
-    @required this.onSavedTask,
-    @required this.onDeleteTask,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) => SingleChildScrollView(
-    child: Row(
-      mainAxisSize: MainAxisSize.min,
-      mainAxisAlignment: MainAxisAlignment.spaceAround,
-      children: [
-        buildSaveButton(),
-        SizedBox(width: 25),
-        buildCancelButton(),
-      ],
-    ),
-  );
-
   Widget buildSaveButton() => Container(
-    height: 40.0,
-    width: 120,
-    decoration: BoxDecoration(
-      color: todoDarkGreen,
-      borderRadius: BorderRadius.circular(30),
-    ),
-    child: FlatButton(
-      onPressed: onSavedTask,
-      child: Center(
-        child: Text(
-          'Save',
-          style: TextStyle(
-              color: Colors.white,
-              fontWeight: FontWeight.w700,
-              fontSize: 16),
+        height: 40.0,
+        width: 100,
+        decoration: BoxDecoration(
+          color: todoDarkGreen,
+          borderRadius: BorderRadius.circular(30),
         ),
-      ),
-    ),
-  );
+        child: FlatButton(
+          onPressed: onSavedTask,
+          child: Center(
+            child: Text(
+              'Save',
+              style: TextStyle(
+                  color: Colors.white,
+                  fontWeight: FontWeight.w700,
+                  fontSize: 16),
+            ),
+          ),
+        ),
+      );
 
   Widget buildCancelButton() => Container(
-    height: 40.0,
-    width: 120,
-    decoration: BoxDecoration(
-      color: todoDarkGreen,
-      borderRadius: BorderRadius.circular(30),
-    ),
-    child: FlatButton(
-      onPressed: onDeleteTask,
-      child: Center(
-        child: Text(
-          'Cancel',
-          style: TextStyle(
-              color: Colors.white,
-              fontWeight: FontWeight.w700,
-              fontSize: 16),
+        height: 40.0,
+        width: 100,
+        decoration: BoxDecoration(
+          color: todoDarkGreen,
+          borderRadius: BorderRadius.circular(30),
         ),
-      ),
-    ),
-  );
+        child: FlatButton(
+          onPressed: onCancel,
+          child: Center(
+            child: Text(
+              'Cancel',
+              style: TextStyle(
+                  color: Colors.white,
+                  fontWeight: FontWeight.w700,
+                  fontSize: 16),
+            ),
+          ),
+        ),
+      );
 }
